requirements.txt
# Core LangChain and RAG dependencies
langchain==0.1.20
langchain-community==0.1.20
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-text-splitters==0.0.1

# Ollama integration
ollama==0.1.9
langchain-ollama==0.1.0

# Vector store and embeddings
chromadb==0.4.24
sentence-transformers==2.7.0
faiss-cpu==1.8.0

# Database connections
sqlalchemy==2.0.30
pyodbc==5.1.0
pymssql==2.2.11
psycopg2-binary==2.9.9
mysql-connector-python==8.4.0

# Configuration and utilities
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
pyyaml==6.0.1
rich==13.7.1
click==8.1.7

# Data processing
pandas==2.2.2
numpy==1.26.4

# Web interface (optional)
streamlit==1.34.0
gradio==4.31.0

# Development and testing
pytest==8.2.1
pytest-asyncio==0.23.7
black==24.4.2
flake8==7.0.0
mypy==1.10.0

# Logging and monitoring
loguru==0.7.2
